requirements.txt
astropy==4.0.1.post1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
comet-git-pure==0.19.16
comet-ml==3.1.12
configobj==5.0.6
cycler==0.10.0
everett==1.0.2
gin-config==0.3.0
idna==2.10
importlib-metadata==1.7.0
jsonschema==3.2.0
kiwisolver==1.2.0
matplotlib==3.2.2
netifaces==0.10.9
numpy==1.19.0
nvidia-ml-py3==7.352.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
tqdm==4.47.0
urllib3==1.25.9
websocket-client==0.57.0
wurlitzer==2.0.0
zipp==3.1.0
